my_blog/requirements.txt
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
defusedxml==0.7.1
Django==2.2
django-allauth==0.45.0
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.13.3
django-notifications-hq==1.6.0
django-password-reset==2.0
django-taggit==1.5.1
idna==3.2
jsonfield==3.1.0
Markdown==3.3.4
oauthlib==3.1.1
Pillow==8.3.2
pycparser==2.20
Pygments==2.10.0
PyJWT==2.2.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
swapper==1.1.2.post1
urllib3==1.26.7
